Added QLearning to Maze
Added q-learning to the project to show how much better it is for this type of task. Even if you still for some reason wanted the LLM to be able to do this task natively, you can just use the q learner to generate the dataset of movements (policy.txt) instead of having the  LLM slap around like a monkey.

I'll separate these later when I have time. This was technically me taking a break ..... (T_T) I have no life.